+++ core/frontend/src/core-frontend.ts
@@ -92,13 +92,10 @@ export * from "./render/MockRender";
 export * from "./render/ParticleCollectionBuilder";
 export * from "./render/Pixel";
 export * from "./render/PolylineArgs";
-export * from "./render/RealityMeshGraphicParams";
 export * from "./render/RealityMeshParams";
 export * from "./render/RenderClipVolume";
 export * from "./render/RenderGraphic";
 export * from "./render/RenderMemory";
-export * from "./render/RenderPlan";
-export * from "./render/RenderPlanarClassifier";
 export * from "./render/RenderSystem";
 export * from "./render/RenderTarget";
 export * from "./render/Scene";
