+++ core/frontend/src/internal/render/webgl/AtmosphereUniforms.ts
@@ -4,7 +4,7 @@
 *--------------------------------------------------------------------------------------------*/
 import { Atmosphere } from "@itwin/core-common";
 import { Matrix3d, Point3d, Transform } from "@itwin/core-geometry";
-import { RenderPlanEllipsoid } from "../../../render/RenderPlan";
+import { RenderPlanEllipsoid } from "../RenderPlan";
 import { WebGLDisposable } from "./Disposable";
 import { Matrix3, Matrix4 } from "./Matrix";
 import { desync, sync, SyncTarget } from "./Sync";
