+++ core/frontend/src/internal/render/webgl/RealityMesh.ts
@@ -11,11 +11,11 @@ import { assert, dispose, disposeArray, UintArray } from "@itwin/core-bentley";
 import { ColorDef, Quantization, RenderTexture } from "@itwin/core-common";
 import { Matrix4d, Range2d, Range3d, Transform, Vector2d } from "@itwin/core-geometry";
 import { GraphicBranch } from "../../../render/GraphicBranch";
-import { RealityMeshGraphicParams } from "../../../render/RealityMeshGraphicParams";
+import { RealityMeshGraphicParams } from "../RealityMeshGraphicParams";
 import { RealityMeshParams } from "../../../render/RealityMeshParams";
 import { RenderGraphic } from "../../../render/RenderGraphic";
 import { RenderMemory } from "../../../render/RenderMemory";
-import { RenderPlanarClassifier } from "../../../render/RenderPlanarClassifier";
+import { RenderPlanarClassifier } from "../RenderPlanarClassifier";
 import { RenderSystem, TerrainTexture } from "../../../render/RenderSystem";
 import { BufferHandle, BufferParameters, QBufferHandle2d, QBufferHandle3d } from "./AttributeBuffers";
 import { AttributeMap } from "./AttributeMap";
