+++ core/frontend/src/internal/render/webgl/Target.ts
@@ -24,8 +24,8 @@ import { AnimationBranchStates } from "../../../render/GraphicBranch";
 import { Pixel } from "../../../render/Pixel";
 import { GraphicList } from "../../../render/RenderGraphic";
 import { RenderMemory } from "../../../render/RenderMemory";
-import { createEmptyRenderPlan, RenderPlan } from "../../../render/RenderPlan";
-import { PlanarClassifierMap, RenderPlanarClassifier } from "../../../render/RenderPlanarClassifier";
+import { createEmptyRenderPlan, RenderPlan } from "../RenderPlan";
+import { PlanarClassifierMap, RenderPlanarClassifier } from "../RenderPlanarClassifier";
 import { RenderTextureDrape, TextureDrapeMap } from "../../../render/RenderSystem";
 import { PrimitiveVisibility, RenderTarget, RenderTargetDebugControl } from "../../../render/RenderTarget";
 import { ScreenSpaceEffectContext } from "../../../render/ScreenSpaceEffectBuilder";
