+++ core/frontend/src/tile/TileDrawArgs.ts
@@ -13,7 +13,7 @@ import { FeatureSymbology } from "../render/FeatureSymbology";
 import { GraphicBranch, GraphicBranchOptions } from "../render/GraphicBranch";
 import { RenderClipVolume } from "../render/RenderClipVolume";
 import { RenderGraphic } from "../render/RenderGraphic";
-import { RenderPlanarClassifier } from "../render/RenderPlanarClassifier";
+import { RenderPlanarClassifier } from "../internal/render/RenderPlanarClassifier";
 import { RenderTextureDrape } from "../render/RenderSystem";
 import { SceneContext } from "../ViewContext";
 import { ViewingSpace } from "../ViewingSpace";
