+++ core/frontend/src/tile/map/MapTile.ts
@@ -12,7 +12,7 @@ import { AxisOrder, BilinearPatch, ClipPlane, ClipPrimitive, ClipShape, ClipVect
 import { IModelApp } from "../../IModelApp";
 import { GraphicBuilder } from "../../render/GraphicBuilder";
 import { RealityMeshParams } from "../../render/RealityMeshParams";
-import { upsampleRealityMeshParams } from "../../render/UpsampleRealityMeshParams";
+import { upsampleRealityMeshParams } from "../../internal/render/UpsampleRealityMeshParams";
 import { RenderGraphic } from "../../render/RenderGraphic";
 import { RenderMemory } from "../../render/RenderMemory";
 import { RenderSystem, RenderTerrainGeometry, TerrainTexture } from "../../render/RenderSystem";
